refactor: `src` 디렉토리 구조로 수정, context 디렉토리 삭제 후 내용을 `iriondaengdaeng/page.tsx`로 이동
