+++ data/bookings.json
@@ -32,5 +32,56 @@
     ],
     "inquiry": "",
     "createdAt": "2024-11-17T05:26:35.747Z"
+  },
+  {
+    "dateTime": {
+      "date": "2024-11-07T15:00:00.000Z",
+      "time": "09:30"
+    },
+    "petInfo": {
+      "petName": "공아지",
+      "weight": "12",
+      "phoneNumber": "1231232131232"
+    },
+    "services": [
+      "hotel",
+      "walking"
+    ],
+    "inquiry": "문의사항123",
+    "createdAt": "2024-11-17T06:13:14.936Z"
+  },
+  {
+    "dateTime": {
+      "date": "2024-11-16T15:00:00.000Z",
+      "time": "11:00"
+    },
+    "petInfo": {
+      "petName": "공아지",
+      "weight": "12",
+      "phoneNumber": "1231232131232"
+    },
+    "services": [
+      "pickup",
+      "walking"
+    ],
+    "inquiry": "문의사항123",
+    "createdAt": "2024-11-17T06:13:41.086Z"
+  },
+  {
+    "dateTime": {
+      "date": "2024-11-14T15:00:00.000Z",
+      "time": "09:00"
+    },
+    "petInfo": {
+      "petName": "강아지",
+      "weight": "12",
+      "phoneNumber": "01076620146"
+    },
+    "services": [
+      "grooming",
+      "walking"
+    ],
+    "inquiry": "문의사항이당~",
+    "createdAt": "2024-11-17T06:19:37.471Z"
   }
 ]
